+++ examples/data-sources/iosxr_extcommunity_opaque_set/data-source.tf
@@ -1,0 +1,3 @@
+data "iosxr_extcommunity_opaque_set" "example" {
+  set_name = "BLUE"
+}