+++ internal/provider/resource_iosxr_router_bgp_neighbor_group_test.go
@@ -19,8 +19,12 @@ func TestAccIosxrRouterBGPNeighborGroup(t *testing.T) {
 					resource.TestCheckResourceAttr("iosxr_router_bgp_neighbor_group.test", "name", "GROUP1"),
 					resource.TestCheckResourceAttr("iosxr_router_bgp_neighbor_group.test", "remote_as", "65001"),
 					resource.TestCheckResourceAttr("iosxr_router_bgp_neighbor_group.test", "update_source", "Loopback0"),
+					resource.TestCheckResourceAttr("iosxr_router_bgp_neighbor_group.test", "bfd_minimum_interval", "3"),
+					resource.TestCheckResourceAttr("iosxr_router_bgp_neighbor_group.test", "bfd_fast_detect", "true"),
 					resource.TestCheckResourceAttr("iosxr_router_bgp_neighbor_group.test", "address_families.0.af_name", "ipv4-labeled-unicast"),
 					resource.TestCheckResourceAttr("iosxr_router_bgp_neighbor_group.test", "address_families.0.soft_reconfiguration_inbound_always", "true"),
+					resource.TestCheckResourceAttr("iosxr_router_bgp_neighbor_group.test", "address_families.0.next_hop_self_inheritance_disable", "true"),
+					resource.TestCheckResourceAttr("iosxr_router_bgp_neighbor_group.test", "address_families.0.route_reflector_client_inheritance_disable", "true"),
 				),
 			},
 			{
@@ -48,9 +52,13 @@ func testAccIosxrRouterBGPNeighborGroupConfig_all() string {
 		name = "GROUP1"
 		remote_as = "65001"
 		update_source = "Loopback0"
+		bfd_minimum_interval = 3
+		bfd_fast_detect = true
 		address_families = [{
 			af_name = "ipv4-labeled-unicast"
 			soft_reconfiguration_inbound_always = true
+			next_hop_self_inheritance_disable = true
+			route_reflector_client_inheritance_disable = true
 		}]
 	}
 	`